requirements.txt
amqp==5.1.1
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
billiard==3.6.4.0
boto3==1.26.87
botocore==1.29.87
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.1.4
django-celery-results==2.4.0
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-decouple==2.1
django-filter==22.1
django-s3-upload==0.4
django-templated-mail==1.1.1
django-tinymce==3.5.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-nested-routers==0.93.4
exceptiongroup==1.1.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.2.4
lxml==4.9.2
MarkupSafe==2.1.1
model-bakery==1.10.1
oauthlib==3.2.2
oscrypto==1.3.0
packaging==23.0
Pillow==9.3.0
pluggy==1.0.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyHanko==0.16.0
pyhanko-certvalidator==0.19.8
PyJWT==2.6.0
PyPDF3==1.0.6
pytest==7.2.1
pytest-django==4.5.2
python-bidi==0.4.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
redis==4.4.2
reportlab==3.6.12
requests==2.28.2
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
tomli==2.0.1
tqdm==4.64.1
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
whitenoise==6.3.0
xhtml2pdf==0.2.8
